Updater: Don't allow stale updates on empty state list
Stale updates will not be added to list on the following conditions:

  -- Device is coming off a fresh wipe or a new Eos install
  -- User removes all entries from updater.

This addresses the issue in which the user installs Eos and gets
a notification that an update is available. However, this update
may be the same or older than currently installed build. Also, if
a user deletes all the entries in the updater, then checks for updates,
updates older than currently installed build will show.

The purpose of an updater is to provide UPDATES. There are other solutions
to provide historical builds.

Change-Id: I0cdf2444d8417c2d0942b0003f0a4954f0e680a9